git-svn-id: http://code.icecube.wisc.edu/svn/projects/dataclasses/releases/B01-13-00@33435 16731396-06f5-0310-8873-f7f720988828

+++ private/dataclasses/I3Time.cxx
@@ -57,7 +57,7 @@ void I3Time::SetModJulianTime(int32_t modJulianDay,
   int32_t secsafteryear = daysafteryear * 3600 * 24 + sec;
   daqTime_ =
     ((int64_t)secsafteryear * ((int64_t)(1e10)))
-    + ((int64_t)ns * ((int64_t)10));
+    + ((int64_t)(ns * 10.));
 }
 
 void I3Time::SetUnixTime(time_t unixTime)
